build(webpack): move rules for ts and sass into single module
